feat(DPE-5176): Substrate agnostic upgrade helpers in their library
